Add a test case that exercises pybind11_protobuf use in parallel.
Tests both with and without the GIL held, and it uncovered an issue in
pybind11 handling of std::string_view when the GIL is not held.

BUG: https://github.com/pybind/pybind11/issues/2765

The bug is that the string_view constructs an object which should be placed on life_support, but due to the underlying mechanism it is not retained in a multi-threaded environment.

PiperOrigin-RevId: 394360131