Ensure that py <-> c++ message descriptors remain live.
Generally removing a python descriptor pool for any proto passed to c++ is unsafe because there may be C++ objects pointing at the deleted pool, and the fast_cpp_proto implementation does not support python weak references, making cleanup impractical.

Replace the weak_ref of a python pool in the py -> c++ mapping with a strong ref that is owned for the life of the process.

If the underlying descriptor pool for a fast_cpp_proto message is not the global
pool, then a python reference to the pool is added for the life of the process.

This fixes a bug where a fast_cpp_proto python protocol buffer is not identified as having the same underlying descriptor database as the equivalent C++ protocol buffer type.

PiperOrigin-RevId: 418663721