Revise PyProto_API resolution.
Internally the PyProto_API is always available, however in the public build it is not, so we cannot unconditionally assume that we can load it and call it.

Introduce nullptr checks and rework several calling locations according to this assumption, and reorder some of the initialization.

Without the PyProto_API, some dynamic message passing fails.

PiperOrigin-RevId: 422072545