Resolve some methods using python MRO (method resolution order) bases.
In some variants of the python protobuf API, the mechanism used to setup
the metaclass for protocol buffer objects leaves the __dict__ in a
state where the default getattr function fails to find the correct method
instances.

In order to work around this, resolve the serialization methods against
the bases in the order specified by type->tp_mro.

Also add another check to try and make sure that PyProto_API use is kosher.

PiperOrigin-RevId: 422606904