Add support for wrapping std::vector<Proto> to WithWrappedProtos.
In the other cases, we introduced a transparent wrapper, WrappedProto<Proto, Kind>, and a type caster. However for containers that doesn't work really well since to convert from std::vector<Wrapper<X>> to std::vector<X> involves a copy/move, even if the wrapper is trivial.

So this is a work around. Since and pybind11 will already delegate a std::vector<T> to the std::vector caster using an inner caster for T, we needed somehow to stop that forwarding. The solution is to rewrite the type as a different low-cost wrapper around std::vector<T>, and then to handle the vector conversion ourselves.

This caster *always* copies values.

PiperOrigin-RevId: 441908466